GPU header cleanup: Avoid including GPUCommon.h in some places
