+++ GPU/D3D11/TextureCacheD3D11.h
@@ -23,7 +23,6 @@
 #include <wrl/client.h>
 
 #include "GPU/GPU.h"
-#include "GPU/GPUCommon.h"
 #include "GPU/Common/TextureCacheCommon.h"
 
 struct VirtualFramebuffer;
