+++ GPU/GLES/TextureCacheGLES.cpp
@@ -28,6 +28,7 @@
 
 #include "GPU/ge_constants.h"
 #include "GPU/GPUState.h"
+#include "GPU/GPUDefinitions.h"
 #include "GPU/GLES/TextureCacheGLES.h"
 #include "GPU/GLES/FramebufferManagerGLES.h"
 #include "GPU/Common/TextureShaderCommon.h"
