+++ GPU/Software/SoftGpu.h
@@ -18,6 +18,7 @@
 #pragma once
 
 #include <cstdint>
+
 #include "GPU/GPUCommon.h"
 #include "GPU/Common/GPUDebugInterface.h"
 #include "Common/GPU/thin3d.h"
